Add files via upload
I have uploaded the move_rectangle script. I am unable to add ROS package because my ubuntu in VirtualBox is not getting enough video Memory to run it properly so I just make a script to run turtle in rectangle by moving it forward then turning it to 90 degree and again moving and repeat this process 2 times.